requirements.txt
click==8.0.3
cycler==0.11.0
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.4
Pillow==8.4.0
pyparsing==3.0.6
pyserial==3.5
python-dateutil==2.8.2
scipy==1.7.2
six==1.16.0
Werkzeug==2.0.2
